+++ xyblog-backEnd/xyblog-admin/src/main/java/cn/xinyuan/blog/admin/sys/service/ISysRoleService.java
@@ -22,5 +22,5 @@ public interface ISysRoleService extends IService<SysRole> {
 
     int updateRole(SysRole role);
 
-    int deleteBatch(Integer[] roleIds);
+    int deleteBatch(Long[] roleIds);
 }
