+++ xyblog-backEnd/xyblog-admin/src/main/java/cn/xinyuan/blog/admin/sys/service/ISysUserRoleService.java
@@ -1,5 +1,6 @@
 package cn.xinyuan.blog.admin.sys.service;
 
+import cn.xinyuan.blog.entity.sys.DO.SysRole;
 import cn.xinyuan.blog.entity.sys.DO.SysUser;
 import cn.xinyuan.blog.entity.sys.DO.SysUserRole;
 import com.baomidou.mybatisplus.extension.service.IService;
@@ -18,7 +19,7 @@ public interface ISysUserRoleService extends IService<SysUserRole> {
 
     int deleteBatch(Long[] userIds);
 
-    List<Integer> queryRoleIdList(Integer userId);
+    List<Long> queryRoleIdList(Long userId);
 
     int insertByUser(SysUser user);
 }
